requirements.txt
numpy~=1.19.2
matplotlib~=3.3.2
scipy~=1.5.2
pandas~=1.1.3
torch~=1.9.0
pyro-ppl
gpytorch~=1.5.0
scikit-learn~=0.23.2
setuptools~=50.3.1